+++ lib/mvp2/auth/auth_pages/auth_page.dart
@@ -9,8 +9,8 @@ import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:homebased_project/mvp2/main/main_pages/widget_tree.dart';
 import 'package:homebased_project/backend/auth_api/auth_service.dart';
 import 'package:homebased_project/mvp2/main/main_components/main_snackbar_widget.dart';
-import 'package:homebased_project/backend/user_profile_api/user_profile_model.dart';
-import 'package:homebased_project/backend/user_profile_api/user_profile_service.dart';
+import 'package:homebased_project/mvp2/profile/profile_data/user_profile_model.dart';
+import 'package:homebased_project/mvp2/profile/profile_data/user_profile_service.dart';
 import 'package:homebased_project/mvp2/auth/auth_pages/reset_password_page.dart';
 import 'package:homebased_project/mvp2/auth/auth_pages/get_reset_password_email_page.dart';
